+++ app/src/main/java/org/astraea/balancer/log/ClusterLogAllocation.java
@@ -1,4 +1,4 @@
-package org.astraea.balancer;
+package org.astraea.balancer.log;
 
 import java.util.ArrayList;
 import java.util.Collection;
