+++ app/src/main/java/org/astraea/balancer/log/IllegalMigrationException.java
@@ -1,4 +1,4 @@
-package org.astraea.balancer;
+package org.astraea.balancer.log;
 
 public class IllegalMigrationException extends RuntimeException {
   public IllegalMigrationException(String s) {
