+++ app/src/main/java/org/astraea/balancer/log/LogPlacement.java
@@ -1,4 +1,4 @@
-package org.astraea.balancer;
+package org.astraea.balancer.log;
 
 import java.util.Optional;
