+++ app/src/test/java/org/astraea/balancer/ClusterLogAllocationTest.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import org.astraea.admin.TopicPartition;
+import org.astraea.balancer.log.ClusterLogAllocation;
+import org.astraea.balancer.log.LogPlacement;
 import org.astraea.cost.ClusterInfoProvider;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
